Начал делать рефактор проекта
